requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2
django-debug-toolbar==1.11
django-extensions==2.1.6
django-rest-framework==0.1.0
djangorestframework==3.9.2
idna==2.8
psycopg2-binary==2.8.1
pytz==2019.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.1
